+++ src/main/java/io/github/astrapi69/crypt/data/key/KeyInfoExtensions.java
@@ -35,37 +35,35 @@ import io.github.astrapi69.crypt.api.key.KeyType;
 import io.github.astrapi69.crypt.data.key.reader.CertificateReader;
 import io.github.astrapi69.crypt.data.key.reader.PrivateKeyReader;
 import io.github.astrapi69.crypt.data.key.reader.PublicKeyReader;
-import io.github.astrapi69.crypt.data.model.KeyModel;
+import io.github.astrapi69.crypt.data.model.KeyInfo;
 import io.github.astrapi69.throwable.RuntimeExceptionDecorator;
 
 /**
- * The class {@link KeyModelExtensions} provides algorithms for transform a given {@link KeyModel}
- * object to the appropriate key objects and given keys to the appropriate {@link KeyModel} object
- * 
- * @deprecated use instead the class {@link KeyInfoExtensions}. Note will be removed in next minor
- *             version
+ * The class {@link KeyInfoExtensions} provides algorithms for transform a given {@link KeyInfo}
+ * object to the appropriate key objects and given keys to the appropriate {@link KeyInfo} object
  */
-public class KeyModelExtensions
+public class KeyInfoExtensions
 {
 
 	/**
 	 * Reads the given byte array with the given algorithm and returns the {@link PrivateKey}
 	 * object.
 	 *
-	 * @param keyModel
+	 * @param keyInfo
 	 *            the info model for create the private key
 	 * @return the {@link PrivateKey} object
 	 */
-	public static PrivateKey toPrivateKey(KeyModel keyModel)
+	public static PrivateKey toPrivateKey(KeyInfo keyInfo)
 	{
-		Objects.requireNonNull(keyModel);
-		if (!keyModel.getKeyType().equals(KeyType.PRIVATE_KEY))
+		Objects.requireNonNull(keyInfo);
+		KeyType keyType = KeyType.toKeyType(keyInfo.getKeyType());
+		if (!keyType.equals(KeyType.PRIVATE_KEY))
 		{
 			throw new RuntimeException(
-				"Given KeyModel:" + keyModel.toString() + "\n is not a private key");
+				"Given KeyInfo:" + keyInfo.toString() + "\n is not a private key");
 		}
 		return RuntimeExceptionDecorator.decorate(
-			() -> PrivateKeyReader.readPrivateKey(keyModel.getEncoded(), keyModel.getAlgorithm()));
+			() -> PrivateKeyReader.readPrivateKey(keyInfo.getEncoded(), keyInfo.getAlgorithm()));
 	}
 
 
@@ -73,81 +71,84 @@ public class KeyModelExtensions
 	 * Reads the given byte array with the given algorithm and returns the {@link PrivateKey}
 	 * object.
 	 *
-	 * @param keyModel
+	 * @param keyInfo
 	 *            the info model for create the private key
 	 * @return the {@link PrivateKey} object
 	 */
-	public static PublicKey toPublicKey(KeyModel keyModel)
+	public static PublicKey toPublicKey(KeyInfo keyInfo)
 	{
-		Objects.requireNonNull(keyModel);
-		if (!keyModel.getKeyType().equals(KeyType.PUBLIC_KEY))
+		Objects.requireNonNull(keyInfo);
+		KeyType keyType = KeyType.toKeyType(keyInfo.getKeyType());
+		if (!keyType.equals(KeyType.PUBLIC_KEY))
 		{
 			throw new RuntimeException(
-				"Given KeyModel:" + keyModel.toString() + "\n is not a public key");
+				"Given KeyInfo:" + keyInfo.toString() + "\n is not a public key");
 		}
 		return RuntimeExceptionDecorator.decorate(
-			() -> PublicKeyReader.readPublicKey(keyModel.getEncoded(), keyModel.getAlgorithm()));
+			() -> PublicKeyReader.readPublicKey(keyInfo.getEncoded(), keyInfo.getAlgorithm()));
 	}
 
 	/**
 	 * Reads the given byte array with the given algorithm and returns the {@link PrivateKey}
 	 * object.
 	 *
-	 * @param keyModel
+	 * @param keyInfo
 	 *            the info model for create the private key
 	 * @return the {@link PrivateKey} object
 	 */
-	public static X509Certificate toX509Certificate(KeyModel keyModel)
+	public static X509Certificate toX509Certificate(KeyInfo keyInfo)
 	{
-		Objects.requireNonNull(keyModel);
-		if (!keyModel.getKeyType().equals(KeyType.CERTIFICATE))
+		Objects.requireNonNull(keyInfo);
+		if (!KeyType.toKeyType(keyInfo.getKeyType()).equals(KeyType.CERTIFICATE))
 		{
 			throw new RuntimeException(
-				"Given KeyModel:" + keyModel.toString() + "\n is not a X509Certificate");
+				"Given KeyInfo:" + keyInfo.toString() + "\n is not a X509Certificate");
 		}
 		return RuntimeExceptionDecorator
-			.decorate(() -> CertificateReader.readCertificate(keyModel.getEncoded()));
+			.decorate(() -> CertificateReader.readCertificate(keyInfo.getEncoded()));
 	}
 
 	/**
-	 * Transforms the given {@link PrivateKey} object to a {@link KeyModel} object
+	 * Transforms the given {@link PrivateKey} object to a {@link KeyInfo} object
 	 * 
 	 * @param privateKey
 	 *            the {@link PrivateKey} object to transform
-	 * @return the {@link KeyModel} object from the given {@link PrivateKey} object
+	 * @return the {@link KeyInfo} object from the given {@link PrivateKey} object
 	 */
-	public static KeyModel toKeyModel(PrivateKey privateKey)
+	public static KeyInfo toKeyInfo(PrivateKey privateKey)
 	{
-		return KeyModel.builder().encoded(privateKey.getEncoded()).keyType(KeyType.PRIVATE_KEY)
-			.algorithm(privateKey.getAlgorithm()).build();
+		return KeyInfo.builder().encoded(privateKey.getEncoded())
+			.keyType(KeyType.PRIVATE_KEY.getDisplayValue()).algorithm(privateKey.getAlgorithm())
+			.build();
 	}
 
 	/**
-	 * Transforms the given {@link PublicKey} object to a {@link KeyModel} object
+	 * Transforms the given {@link PublicKey} object to a {@link KeyInfo} object
 	 * 
 	 * @param publicKey
 	 *            the {@link PublicKey} object to transform
-	 * @return the {@link KeyModel} object from the given {@link PublicKey} object
+	 * @return the {@link KeyInfo} object from the given {@link PublicKey} object
 	 */
-	public static KeyModel toKeyModel(PublicKey publicKey)
+	public static KeyInfo toKeyInfo(PublicKey publicKey)
 	{
-		return KeyModel.builder().encoded(publicKey.getEncoded()).keyType(KeyType.PUBLIC_KEY)
-			.algorithm(publicKey.getAlgorithm()).build();
+		return KeyInfo.builder().encoded(publicKey.getEncoded())
+			.keyType(KeyType.PUBLIC_KEY.getDisplayValue()).algorithm(publicKey.getAlgorithm())
+			.build();
 	}
 
 	/**
-	 * Transforms the given {@link Certificate} object to a {@link KeyModel} object
+	 * Transforms the given {@link Certificate} object to a {@link KeyInfo} object
 	 * 
 	 * @param certificate
 	 *            the {@link Certificate} object to transform
-	 * @return the {@link KeyModel} object from the given {@link Certificate} object
+	 * @return the {@link KeyInfo} object from the given {@link Certificate} object
 	 * @throws CertificateEncodingException
 	 *             is thrown if an encoding error occurs
 	 */
-	public static KeyModel toKeyModel(Certificate certificate) throws CertificateEncodingException
+	public static KeyInfo toKeyInfo(Certificate certificate) throws CertificateEncodingException
 	{
-		return KeyModel.builder().encoded(CertificateExtensions.getEncoded(certificate))
-			.keyType(KeyType.CERTIFICATE).build();
+		return KeyInfo.builder().encoded(CertificateExtensions.getEncoded(certificate))
+			.keyType(KeyType.CERTIFICATE.getDisplayValue()).build();
 	}
 
 }