+++ src/main/java/io/github/astrapi69/crypt/data/model/KeyInfo.java
@@ -24,7 +24,6 @@
  */
 package io.github.astrapi69.crypt.data.model;
 
-import io.github.astrapi69.crypt.api.key.KeyType;
 import lombok.AccessLevel;
 import lombok.Data;
 import lombok.NonNull;
@@ -34,20 +33,29 @@ import lombok.experimental.SuperBuilder;
 
 /**
  * Data class representing key information.
- * 
- * @deprecated use instead the class {@link io.github.astrapi69.crypt.data.model.KeyInfo}. Note will
- *             be removed in next minor version
  */
 @Data
 @RequiredArgsConstructor
 @SuperBuilder(toBuilder = true)
 @FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
-public class KeyModel
+public class KeyInfo
 {
+
+	/**
+	 * The type of the key.
+	 */
 	@NonNull
-	KeyType keyType;
+	String keyType;
+
+	/**
+	 * The encoded key.
+	 */
 	@NonNull
 	byte[] encoded;
+
+	/**
+	 * The algorithm of the key.
+	 */
 	@NonNull
 	String algorithm;
 }