+++ src/main/java/io/github/astrapi69/crypt/data/model/KeyPairInfo.java
@@ -1,6 +1,33 @@
+/**
+ * The MIT License
+ *
+ * Copyright (C) 2015 Asterios Raptis
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+ * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+ * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+ * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
 package io.github.astrapi69.crypt.data.model;
 
+import java.security.InvalidAlgorithmParameterException;
 import java.security.KeyPair;
+import java.security.NoSuchAlgorithmException;
+import java.security.NoSuchProviderException;
 
 import io.github.astrapi69.crypt.data.factory.KeyPairFactory;
 import lombok.AccessLevel;
@@ -10,35 +37,57 @@ import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import lombok.experimental.SuperBuilder;
 
+/**
+ * Data class representing key pair information.
+ */
 @Data
 @RequiredArgsConstructor
 @SuperBuilder(toBuilder = true)
 @FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
 public class KeyPairInfo
 {
+
+	/**
+	 * The name of the EC named curve parameter specification.
+	 */
 	String eCNamedCurveParameterSpecName;
+
+	/**
+	 * The algorithm used for the key pair.
+	 */
 	@NonNull
 	String algorithm;
+
+	/**
+	 * The provider of the key pair.
+	 */
 	String provider;
+
+	/**
+	 * The key size for the key pair.
+	 */
 	int keySize;
 
-
-	public static KeyPair newKeyPair(KeyPairInfo keyPairInfo) throws Exception
+	/**
+	 * Factory method to create a new {@link KeyPair} object from the given {@link KeyPairInfo}
+	 * object.
+	 *
+	 * @param keyPairInfo
+	 *            the {@link KeyPairInfo} object containing the key pair details.
+	 * @return the newly created {@link KeyPair} object.
+	 *
+	 * @throws InvalidAlgorithmParameterException
+	 *             is thrown if initialization of the cipher object fails
+	 * @throws NoSuchAlgorithmException
+	 *             is thrown if no Provider supports a KeyPairGeneratorSpi implementation for the
+	 *             specified algorithm
+	 * @throws NoSuchProviderException
+	 *             is thrown if the specified provider is not registered in the security provider
+	 *             list
+	 */
+	public static KeyPair newKeyPair(KeyPairInfo keyPairInfo)
+		throws InvalidAlgorithmParameterException, NoSuchAlgorithmException, NoSuchProviderException
 	{
-		KeyPair keyPair;
-		if (keyPairInfo.getECNamedCurveParameterSpecName() != null
-			&& keyPairInfo.getProvider() != null)
-		{
-			keyPair = KeyPairFactory.newKeyPair(keyPairInfo.getECNamedCurveParameterSpecName(),
-				keyPairInfo.getAlgorithm(), keyPairInfo.getProvider());
-			return keyPair;
-		}
-		if (keyPairInfo.getECNamedCurveParameterSpecName() != null)
-		{
-			keyPair = KeyPairFactory.newKeyPair(keyPairInfo.getECNamedCurveParameterSpecName(),
-				keyPairInfo.getAlgorithm(), "BC");
-			return keyPair;
-		}
-		return KeyPairFactory.newKeyPair(keyPairInfo.getAlgorithm(), keyPairInfo.getKeySize());
+		return KeyPairFactory.newKeyPair(keyPairInfo);
 	}
 }