+++ src/main/java/io/github/astrapi69/crypt/data/model/PemKeyModel.java
@@ -32,16 +32,34 @@ import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import lombok.experimental.SuperBuilder;
 
+/**
+ * Data class representing a PEM-encoded key model.
+ * 
+ * @deprecated use instead the class {@link io.github.astrapi69.crypt.data.model.PemKeyInfo}. Note
+ *             will be removed in next minor version
+ */
 @Data
 @RequiredArgsConstructor
 @SuperBuilder(toBuilder = true)
 @FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
 public class PemKeyModel
 {
+
+	/**
+	 * The type of the PEM key.
+	 */
 	@NonNull
 	PemType keyType;
+
+	/**
+	 * The PEM-encoded key.
+	 */
 	@NonNull
 	String pemEncoded;
+
+	/**
+	 * The algorithm used for the key.
+	 */
 	@NonNull
 	String algorithm;
 }