+++ src/main/java/io/github/astrapi69/crypt/data/model/X509CertificateV3Info.java
@@ -24,7 +24,8 @@
  */
 package io.github.astrapi69.crypt.data.model;
 
-import io.github.astrapi69.crypt.api.key.KeyType;
+import java.math.BigInteger;
+
 import lombok.AccessLevel;
 import lombok.Data;
 import lombok.NonNull;
@@ -32,22 +33,12 @@ import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import lombok.experimental.SuperBuilder;
 
-/**
- * Data class representing key information.
- * 
- * @deprecated use instead the class {@link io.github.astrapi69.crypt.data.model.KeyInfo}. Note will
- *             be removed in next minor version
- */
 @Data
 @RequiredArgsConstructor
 @SuperBuilder(toBuilder = true)
 @FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
-public class KeyModel
+public class X509CertificateV3Info
 {
-	@NonNull
-	KeyType keyType;
-	@NonNull
-	byte[] encoded;
-	@NonNull
-	String algorithm;
+	X509CertificateV1Info certificateV1Info;
+	ExtensionInfo[] extensions;
 }